+++ NEWS.md
@@ -1,10 +1,12 @@
 
-# LikertMakeR 0.4.0 (2024-11-10)
+# LikertMakeR 0.4.0 (2024-11-15)
 
 ## target Cronbach's Alpha added to makeItemsScale() function 
 
 generated scale items now defined by a target Cronbach's Alpha, 
-as well as by variance within each scale item
+as well as by variance within each scale item.
+This latest version adds a little randomness to the selection of 
+candidate row vectors.